% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_circle_markers_county.R
\name{add_circle_markers_county}
\alias{add_circle_markers_county}
\title{Add circle markers to leaflet map}
\usage{
add_circle_markers_county(
  map,
  dat = NULL,
  pal_domain = NULL,
  county_pal,
  size,
  alpha = 0.65,
  group = NULL
)
}
\arguments{
\item{map}{A map widget object created from leaflet().}

\item{dat}{Data frame with 3 columns: \code{COUNTY}, \code{LONGITUDE}, and
\code{LATITUDE}. Optional additional column for scaling \code{size}. The
default inherits \code{dat} from \code{leaflet()}.}

\item{pal_domain}{The possible values that can be mapped (e.g., counties).
The default is the unique values of \code{COUNTY} in \code{dat}. For layers
that do not include all counties (e.g., DO in mg/L), a character vector of
all can be used so that the colours are consistent between layers.}

\item{county_pal}{Colour palette to assign to counties.}

\item{size}{Radius of the circle markers. Can be numeric or based on a column
in \code{dat}, e.g., \code{~n_prop}.}

\item{alpha}{Transparency of markers.}

\item{group}{Name of the group the layer belongs to.}
}
\value{
a new map object
}
\description{
Add circle markers to leaflet map
}
