% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_circle_markers_county.R
\name{add_county_fill_legend}
\alias{add_county_fill_legend}
\title{Add legend for county colours to leaflet plot}
\usage{
add_county_fill_legend(map, dat = NULL, county_pal, alpha = 0.65)
}
\arguments{
\item{map}{A map widget object created from leaflet().}

\item{dat}{Data frame with 3 columns: \code{COUNTY}, \code{LONGITUDE}, and
\code{LATITUDE}. Optional additional column for scaling \code{size}. The
default inherits \code{dat} from \code{leaflet()}.}

\item{county_pal}{Colour palette to assign to counties.}

\item{alpha}{Transparency of markers.}
}
\value{
a new map object
}
\description{
Add legend for county colours to leaflet plot
}
