% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county_colour_palette.R
\name{get_county_colour_palette}
\alias{get_county_colour_palette}
\title{Generate consistent colour palette for counties}
\usage{
get_county_colour_palette(n_col)
}
\arguments{
\item{n_col}{Number of colours to generate (number of counties).}
}
\value{
n_col hex codes
}
\description{
Generate consistent colour palette for counties
}
