% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram_user_thresholds.R
\name{plot_histogram_user_thresholds}
\alias{plot_histogram_user_thresholds}
\title{Plot distribution of observations}
\usage{
plot_histogram_user_thresholds(
  dat,
  hist_col,
  binwidth = NULL,
  user_min = -Inf,
  user_max = Inf,
  fill_col = NULL,
  pal = "grey50",
  legend_name = NULL
)
}
\arguments{
\item{dat}{Data frame with at least one column.}

\item{hist_col}{Name of the column to plot (in quotes).}

\item{binwidth}{Width of the bins. Passed to \code{geom_histogram()}.}

\item{user_min}{user_min threshold}

\item{user_max}{user_max threshold}

\item{fill_col}{Name of the column used to for the fill colour. If
\code{fill_col = NULL}, the default fill colour will be used.}

\item{pal}{Colour palette used for the fill.}

\item{legend_name}{Name of the legend. Default is \code{fill_col}. There will
not be a legend if \code{fill_col = NULL}.}
}
\value{
Returns a gpplot object.
}
\description{
Plot distribution of observations
}
