% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_sd.R
\name{plot_mean_sd_county}
\alias{plot_mean_sd_county}
\title{Plot the mean and +/- 1 standard deviation for each county}
\usage{
plot_mean_sd_county(dat, ordered = TRUE, text_size = 14)
}
\arguments{
\item{dat}{Data frame with at least four columns: \code{county}, \code{mean},
\code{stdev}, and \code{variable}.}

\item{ordered}{If TRUE, orders the counties based on the mean. If FALSE,
orders counties alphabetically.}

\item{text_size}{Numeric value for the size of the text.}
}
\value{
ggplot object.
}
\description{
Plot the mean and +/- 1 standard deviation for each county
}
