% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_obs.R
\name{plot_n_obs_county}
\alias{plot_n_obs_county}
\title{Plot the number of observations for each county}
\usage{
plot_n_obs_county(dat, text_size = 14)
}
\arguments{
\item{dat}{Data frame with at least three columns: \code{n}, \code{county},
\code{units}.}

\item{text_size}{Numeric value for the size of the text.}
}
\value{
ggplot object.
}
\description{
Plot the number of observations for each county
}
