% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_obs.R
\name{plot_n_obs_depth}
\alias{plot_n_obs_depth}
\title{Plot the number of observations for each depth}
\usage{
plot_n_obs_depth(dat, facet_county = TRUE, text_size = 14)
}
\arguments{
\item{dat}{Data frame with at least two columns: \code{n}, \code{depth}.}

\item{facet_county}{If TRUE, plot will be faceted by county.}

\item{text_size}{Numeric value for the size of the text.}
}
\value{
ggplot object
}
\description{
Plot the number of observations for each depth
}
