% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_obs.R
\name{plot_n_obs_month}
\alias{plot_n_obs_month}
\title{Plot the number of observations for each month}
\usage{
plot_n_obs_month(dat, facet_county = TRUE, text_size = 14, fill = "#AE1759FF")
}
\arguments{
\item{dat}{Data frame with at least three columns: \code{n}, \code{month},
and optionally \code{county}.}

\item{facet_county}{If TRUE, plot will be faceted by county.}

\item{text_size}{Numeric value for the size of the text.}

\item{fill}{Hex code. Colour to fill bars.}
}
\value{
ggplot object.
}
\description{
Plot the number of observations for each month
}
