% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_n_user_flags.R
\name{summarise_n_user_flags}
\alias{summarise_n_user_flags}
\title{Calculate percent of observations flagged by user thresholds}
\usage{
summarise_n_user_flags(dat, user_thresh, group)
}
\arguments{
\item{dat}{Data frame with a least one column: \code{VALUE}.}

\item{user_thresh}{Data frame with columns \code{mean}, \code{stdev},
\code{user_min} and \code{user_max}.}

\item{group}{Identifier for how the thresholds were calculated}
}
\description{
Calculate percent of observations flagged by user thresholds
}
