% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot distrbution of observations}
\usage{
plot_histogram(dat, hist_col, binwidth = NULL)
}
\arguments{
\item{dat}{Data frame with at least one column.}

\item{hist_col}{Name of the column to plot (in quotes).}

\item{binwidth}{Width of the bins. Passed to \code{geom_histogram()}.}
}
\value{
Returns a gpplot object.
}
\description{
Plot distrbution of observations
}
